+++ mex/editor/client/src/app/fast-track/fieldset/fieldset.ts
@@ -1,10 +1,46 @@
-import { Component, computed, HostBinding, input } from "@angular/core";
+import type { PipeTransform } from "@angular/core";
+import { Component, computed, HostBinding, inject, input, Pipe } from "@angular/core";
+import type { FieldState, ValidationError } from "@angular/forms/signals";
 import { type FieldTree } from "@angular/forms/signals";
-import { TranslocoPipe } from "@jsverse/transloco";
+import { TranslocoPipe, TranslocoService } from "@jsverse/transloco";
+
+import type { FieldCategory } from "../../shared/models";
+
+@Pipe({
+  name: "fieldErrorLabel",
+})
+/**
+ * Transforms field errors to human readable text.
+ */
+class FieldErrorLabel implements PipeTransform {
+  private transloco = inject(TranslocoService);
+  transform(
+    value: ValidationError.WithFieldTree,
+    state: FieldState<unknown>,
+    originalLabel: string,
+  ) {
+    const errorKey = value.fieldTree().keyInParent();
+    const isInnerError = errorKey !== state.keyInParent();
+    if (isInnerError) {
+      if (state.value() instanceof Array) {
+        return this.transloco.translate("validation.innerArrayError", {
+          field: originalLabel,
+          position: (typeof errorKey == "string" ? parseInt(errorKey) : errorKey) + 1,
+        });
+      }
+      return this.transloco.translate("validation.innerObjectError", {
+        field: originalLabel,
+        property: errorKey,
+      });
+    }
+
+    return originalLabel;
+  }
+}
 
 @Component({
   selector: "mex-fieldset",
-  imports: [TranslocoPipe],
+  imports: [FieldErrorLabel, TranslocoPipe],
   templateUrl: "./fieldset.html",
   styleUrl: "./fieldset.scss",
 })
@@ -27,7 +63,7 @@ export class Fieldset<T> {
   descriptionKey = input<string>();
 
   showCategoryLabel = input(true);
-  category = input<"required" | "optional" | "recommended">("optional");
+  category = input.required<FieldCategory>();
   categoryKey = computed(() => `categories.${this.category()}`);
 
   showErrorWithoutTouch = input(false);