+++ CHANGELOG.md
@@ -8,8 +8,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- code generator to create zod schemas based on pydantic models
+- datepicker component to unify usage and handle parse errors
+- field-category pipe to determine field-category based on zod schema definition
 
 ### Changes
+- models are now defined by zod schemas
+- fasttrack-resource form with validation
+- fieldsets category is now required and no longer optional
 
 - set exact version for neo4j:2026.07.1